activelife-backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
google-generativeai==0.3.1
cryptography==41.0.7
python-dotenv==1.0.1